AST

Oolong generates an abstract syntax tree of tensor operations. Each operation on tensors is translated to an operation node.

using TensorDSL
A = VariableTensor("A")
B = VariableTensor("B")
A + 3*B

# output

AddOperation{0}
    VariableTensor{0}, A, shape: ()
    OuterProductOperation{0}
        ConstantTensor{Int64,0}, 3, shape: ()
        VariableTensor{0}, B, shape: ()

This tree can then be traversed using traversal-function. Examples of this include togem(node::Node) and differentiate(A::AbstractTensor{0}, x::VariableTensor{0}).

The function assign(node::Node, pair::Pair{VariableTensor{rank}, T}) where {rank, T<:AbstractTensor{rank}} can be used to assign values for the variables in the tree.

TensorDSL.Node — Type

Node

Abstract type for nodes in the AST tree.

TensorDSL.traversal — Method

traversal(node, pretraversalfn::Function, visitfn::Function,
    pretraversalfnargs::Union{Any, Nothing}, visitfnargs::Union{Any, Nothing},
    visited::Union{Dict, Nothing}, posttraversal=false)

Traverses an AST, calling visitfn on the nodes of the tree.

Arguments

node: the root of the tree to traverse. Assumes that each node has children

attribute.

pretraversalfn::Function A function to be executed before the traversal

visitfn::Function The function that is executed on each visit. If

posttraversal is false, and visitfnargs nothing, the function must have signature visitfn(visited, node, new_children...). If visitfnargs is not nothing, the signature is visitfn(visited, node, visitfnargs, new_children...). If posttraversal is true, the signature must bevisitfn(node, fn), orvisitfn(node, visitfn, fn), wherefn` is a function that calls traversal on a node.

pretraversalfnargs::Union{Any, Nothing} Arguments for the pretraversalfn.

visitfnargs::Union{Any, Nothing} Arguments for traversalfn.

visited::Union{Dict, Nothing} If not nothing, a dictionary with the results

of calling traversalfn on different nothing. If nothing, this dictionary is created in traversal.

posttraversal=false Determines the way traversal traverses each node. If

posttraversal=false, the traversal first processes the children of a node, and then the parent node (example: togem(node::Node)). If posttraversal=true, the visitfn first processes the parent and then calls visitfn on its children if needed (example: differentiate(A::AbstractTensor{0}, x::VariableTensor{0})).

TensorDSL.assign — Method

assign(node::Node, pair::Pair{VariableTensor{rank}, T}) where {rank, T<:AbstractTensor{rank}}
assign(node::Node, pair::Pair{T, N}) where {T<:AbstractTensor{0}, N<:Number}
assign(node::Node, pair::Pair{T1, T2}) where {T1<:AbstractVectorSpace, T2<:AbstractVectorSpace}
assign(node::Node, pair::Pair{FreeIndex{T}, Union{Int, FixedIndex{T}}}) where {T<:AbstractVectorSpace}

Create a new node from node, where every value first(pair) is replaced with last(pair). Can be used on tensors, indices and vector spaces.