Compilation

Oolong is compiled using FireDrake's infrastructure. The infrastructure is mostly in Python. The Pycall-module is used to call Python from Julia.

First Oolong-nodes can be translated to GEM, which is a part of TSFC. The function togem(node::Node) translates a given Oolong-node to GEM. The GEM is then futher compiled to Impero and from there to Loopy. Loopy is then compiled to C-code and executed using PyOp2.

In most cases the user does not need to do this directly. They can create a PyOp2 Kernel and execute it with the function execute(knl::Kernel, variables::Dict{String, T}) where T. If they want, they can also execute the node directly with the execute(knl::Kernel, variables::Dict{String, T}) where T-function.

TensorDSL.Kernel — Type

Kernel(node::Node)

Creates a PyOp2 kernel from a given node. This can be executed repeatedly.

TensorDSL.execute — Method

execute(expr::Union{Node, Kernel}, variables::Vararg{Pair})

Execute the Kernel knl or Node with the variable-assignments in variables.

Example

julia> A = VariableTensor("A", RnSpace(3))
julia> execute(A[1], "A"=>[1.0, 2.0, 3.0])
1-element Array{Float64,1}:
 1.0

TensorDSL.execute — Method

execute(expr::Union{Kernel, Node}, variables::Dict{String, T}) where T

Execute the Kernel knl with the variable-assignments in variables.

Example

julia> A = VariableTensor("A", RnSpace(3))
julia> knl = Kernel(A[1])
julia> execute(knl, Dict("A"=>[1.0, 2.0, 3.0]))
1-element Array{Float64,1}:
 1.0

TensorDSL.togem — Method

togem(node::Node)

Translates node into GEM.