Indices

Indices in Oolong are tied to particular AbstractVectorSpace. When creating an index, one must also specify the vector space they are indexing.

Indices on their own are not nodes in AST. Instead, a tuple of indices can be stored in an Indices-node. In most cases the user does not need to create this, as the index-operations create their own Indices-objects.

TensorDSL.FixedIndex — Type

FixedIndex(V::T, value::Int) where {T<:AbstractVectorSpace}

An index whose value is known.

TensorDSL.FreeIndex — Type

FreeIndex(V::T, name::String, id=0) where {T<:AbstractVectorSpace}

An index whose value is not known. Identified by its name. There should be no need to set the id by the user.

TensorDSL.Index — Type

Index

Abstract index supertype.

TensorDSL.Indices — Type

Indices(indices::Vararg{Index})

A node that represents a tuple of indices.

Base.adjoint — Method

Base.adjoint(i::Index)

Creates an index with the same name/value in the dual space.

TensorDSL.toindex — Method

toindex(V::AbstractVectorSpace, i::Union{String, Int, Index})

Given a vectorspace V, creates an index of the string or integer i. If i is an index, returns i.