Base.:+ — Method

Base.:+(nodes::Vararg{Node})

Creates an addoperation whose children are the nodes.

TensorDSL.:⊗ — Method

⊗(A::AbstractTensor, B::AbstractTensor)

Returns an OuterProductOperation with A and B as its children.

Base.:* — Method

Base.:*(x::Scalar, A::AbstractTensor)

Shorthand for multiplying a tensor by a scalar. If multiplying a number by a variable, the number must be first.

Base.:- — Method

Base.:-(A::AbstractTensor)

Unary minus. Creates an OuterProductOperation between ConstantTensor(-1) and A.

Base.:- — Method

Base.:-(A::AbstractTensor{rank}, B::AbstractTensor{rank}) where rank

Binary minus. Creates an AddOperation of A and -1*B.

Base.:^ — Method

Base.:^(x::AbstractTensor{0}, y::AbstractTensor{0})

Raises the scalar-shaped tensor x to the scalar-shaped power y. Returns a PowerOperation.

Base.:^ — Method

Base.:^(x::AbstractTensor{0}, y::Number)

Raises the scalar-shaped tensor x to the power y. Returns a PowerOperation.

Base.sqrt — Method

Base.sqrt(x::AbstractTensor{0})

Returns the square root of x as a PowerOperation.

Base.:/ — Method

Base.:/(A::AbstractTensor, y::AbstractTensor{0})

Creates a DivisionOperation where A is divived by y. y must be scalar-shaped.

Base.:/ — Method

Base.:/(y::Number, A::AbstractTensor{0})

Creates a DivisionOperation where y is divided by A, where A must be a scalar.

TensorDSL.componenttensor — Method

componenttensor(A::AbstractTensor, indices::Vararg{Index})

Creates a component tensor of A over indices. Indices must be a subset of the free indices of A.

Base.getindex — Method

Base.getindex(A::AbstractTensor, ys::Vararg{Index})

Creates an IndexingOperation of A indexed by ys. If every dimension of A is not indexed, creates a ComponentTensorOperation over the unindexed dimensions.

Base.getindex — Method

Base.getindex(A::AbstractTensor, ys::Vararg{Union{String, Int, Index}})

A convenience function that allows indexing a tensor by an integer. The function creates a corresponding FixedIndex object in the appropiate vector space.

Technically the function also does the same for a string, but this is not recommended, because the syntax makes no distinction between upper and lower indices.

Base.transpose — Method

Base.transpose(A::AbstractTensor)

Reverses the shape of A. Does this by indexing A and then creating a componenttensor with the indices reversed.

TensorDSL.trace — Method

trace(A::AbstractTensor{2})

Returns the trace of a matrix of the shape (V, V').