TensorDSL.AbstractTensor — Type

AbstractTensor{rank}

A type for all the tensors and tensor like objects.

The rank means the number of vector spaces in the shape. For example AbstractTensor{0} is a scalar, and AbstractTensor{2} a matrix.

TensorDSL.TerminalTensor — Type

TerminalTensor{rank}

A type for the tensors which are terminal nodes, i.e not created by operations.

TensorDSL.Scalar — Constant

A union type for everything that can be treated as a scalar. Includes tensors of rank 0 and numbers.

TensorDSL.VariableTensor — Type

VariableTensor(name::String, shape::Vararg{AbstractVectorSpace})

Tensors of which we only know in which vector spaces their indices are. The rank is the lenght of the shape.

TensorDSL.Tensor — Type

Tensor(x::Array{T}, Vs::Vararg{AbstractVectorSpace}) where T

Tensor{T, length(VS)} created from a multidimensional array.

TensorDSL.Tensor — Method

Tensor(x::T) where (T <: Scalar)

Turns a non-tensor scalar into a tensor. A tensor scalar does not change.

TensorDSL.ZeroTensor — Type

ZeroTensor(Vs::Vararg{AbstractVectorSpace})

Symbolic zero tensor of the shape Vs.

TensorDSL.ConstantTensor — Type

ConstantTensor(value::T, Vs::Vararg{AbstractVectorSpace}) where (T <: Scalar)

A tensor where every entry is of the same value. Creates ConstantTensor{T, length(Vs)}. If value == 0, creates a ZeroTensor.