Vector Spaces

The shape of Tensors is a represented as a tuple of vector spaces.

Vector spaces with or without a known number of dimensions. (Note: vector spaces with unknow dimension cannot be compiled to GEM.)

V4 = VectorSpace(4)
V = VectorSpace()

Dual spaces

julia> typeof(V4')
DualVectorSpace

Real vector spaces

R2 = RnSpace(2)

TensorDSL.AbstractVectorSpace — Type

Abstract vector space type.

TensorDSL.DualVectorSpace — Type

DualVectorSpace(vectorspace::AbstractVectorSpace)

Dual of a given vector space.

TensorDSL.RnSpace — Type

RnSpace(dim::Int)

Vector space corresponding to R^n for some n.

TensorDSL.VectorSpace — Type

VectorSpace(dim::Union{Int, Nothing})

A vector space with optional dimension. Attribute id distinguishes between different vector spaces of the same dimension. User should not need to set id.

Base.adjoint — Method

Base.adjoint(V::AbstractVectorSpace)

Returns the dual(V).

TensorDSL.dim — Method

dim(V::AbstractVectorSpace)

Returns the dimension V.

TensorDSL.dual — Method

dual(V::AbstractVectorSpace)

Returns the dual of V.