Tensors

The tensor product of two vector spaces $E$ and $F$ is a new vector space, denoted by $E⊗F$. It is spanned by the elements $e⊗f$ where $e \in E= e_i \phi_i$ and $f \in F= f_j \psi_j$ then $g \in E ⊗ F= g_{ij} \phi_i ⊗ \psi_j$.

Tensors are these objects in the vector spaces created by tensors products

Mathematically, tensor $T$ is $T \in E_1⊗…⊗E_p⊗F_1^∗⊗…⊗F_q^∗$ (or a permutation of these), where $E_i$ and $F_j^∗$ are vector and dual spaces, respectively Programmatically this means that tensors objects with shape and data. Shape describes the tensor space $E_1⊗…⊗E_p⊗F_1^∗⊗…⊗F_q^∗$ and data describes the vector within this tensor space.

There are two main types of tensors in Oolong: tensors created from multidimensional arrays:

using TensorDSL
R2 = RnSpace(2)
A = Tensor([1 2 ; 3 4], R2, R2)

# output

Tensor{Int64,2}, [1 2; 3 4], shape: R^2⊗ R^2

and tensors with unknown value

using TensorDSL
R2 = RnSpace(2)
B = VariableTensor("B", R2, R2)

# output

VariableTensor{2}, B, shape: R^2⊗ R^2

Zero and and identity tensors have their own types.

julia> Z = ZeroTensor(RnSpace(3))
ZeroTensor{1}, shape: R^3
julia> I = DeltaTensor(RnSpace(3), RnSpace(3))
DeltaTensor{2}, shape: R^3⊗ R^3

There is also a type for tensors where all entries have the same value.

julia> C = ConstantTensor(3)
ConstantTensor{Int64,0}, 3, shape: ()

TensorDSL.Scalar — Constant

A union type for everything that can be treated as a scalar. Includes tensors of rank 0 and numbers.

TensorDSL.AbstractTensor — Type

AbstractTensor{rank} <: Node

A type for all the tensors and tensor like objects.

The rank means the number of vector spaces in the shape. For example AbstractTensor{0} is a scalar, and AbstractTensor{2} a matrix.

TensorDSL.ConstantTensor — Type

ConstantTensor(value::T, Vs::Vararg{AbstractVectorSpace}) where (T <: Scalar)

A tensor where every entry is of the same value. Creates ConstantTensor{T, length(Vs)}. If value == 0, creates a ZeroTensor.

TensorDSL.DeltaTensor — Type

DeltaTensor(Vs::Vararg{AbstractVectorSpace})

Symbolic delta tensor. Each vector space in Vs must be either equal to other vector spaces or their duals.

TensorDSL.Tensor — Type

Tensor(x::Array{T}, Vs::Vararg{AbstractVectorSpace}) where T

Tensor{T, length(VS)} created from a multidimensional array.

TensorDSL.Tensor — Method

Tensor(x::T) where (T <: Scalar)

Turns a non-tensor scalar into a tensor. A tensor scalar does not change.

TensorDSL.TerminalTensor — Type

TerminalTensor{rank} <: AbstractTensor{rank}

A type for the tensors which are terminal nodes, i.e not created by operations.

TensorDSL.VariableTensor — Type

VariableTensor(name::String, shape::Vararg{AbstractVectorSpace})

Tensors of which we only know in which vector spaces their indices are. The rank is the lenght of the shape.

TensorDSL.ZeroTensor — Type

ZeroTensor(Vs::Vararg{AbstractVectorSpace})

Symbolic zero tensor of the shape Vs.