Operations

Tensor operations

Most basic tensor operations are represented in Oolong.

Tensors of the same shape can be added together.

using TensorDSL
R2 = RnSpace(2)
A = Tensor([1 2; 3 4], R2, R2)
B = VariableTensor("B", R2, R2)
A + B
# output
AddOperation{2}
    Tensor{Int64,2}, [1 2; 3 4], shape: R^2⊗ R^2
    VariableTensor{2}, B, shape: R^2⊗ R^2

Outer product

using TensorDSL
R2 = RnSpace(2)
A = Tensor([1 2; 3 4], R2, R2)
C = ZeroTensor(R2)
A ⊗ C

# output
OuterProductOperation{3}
    Tensor{Int64,2}, [1 2; 3 4], shape: R^2⊗ R^2
    ZeroTensor{1}, shape: R^2

Tensors can be indexed using both FreeIndex objects and integers.

using TensorDSL
V = VectorSpace()
x = FreeIndex(V, "x")
A = VariableTensor("A", VectorSpace(3), V)
A[1, x]
# output
IndexingOperation{0}
    VariableTensor{2}, A, shape: V3_2⊗ V_1
    (1, x)

Component tensor is an operation where a free index of a tensor expression is "rolled out". This turn a free index into a shape in the tensor expression.

using TensorDSL
V = VectorSpace()
x = FreeIndex(V, "x")
A = VariableTensor("A", V', VectorSpace(3))
componenttensor(A[x', 1], x')
# output
ComponentTensorOperation{1}
    IndexingOperation{0}
        VariableTensor{2}, A, shape: V_3*⊗ V3_4
        (x*, 1)
    (x*,)

Tensor contraction is invoked whenever the expression is indexed with an upper and lower index.

using TensorDSL
V = VectorSpace()
x = FreeIndex(V, "x")
A = VariableTensor("A", V)
B = VariableTensor("B", V')
A[x] ⊗ B[x']
# output
IndexSumOperation{0}
    OuterProductOperation{0}
        IndexingOperation{0}
            VariableTensor{1}, A, shape: V_5
            (x,)
        IndexingOperation{0}
            VariableTensor{1}, B, shape: V_5*
            (x*,)
    (x,)

Functions

Base.:* — Method

Base.:*(x::Scalar, A::AbstractTensor)

Shorthand for multiplying a tensor by a scalar. If multiplying a number by a variable, the number must be first.

Base.:+ — Method

Base.:+(nodes::Vararg{Node})

Creates an AddOperation whose children are the nodes. The shapes of the nodes must match.

Base.:- — Method

Base.:-(A::AbstractTensor)

Unary minus. Creates an OuterProductOperation between ConstantTensor(-1) and A.

Base.:- — Method

Base.:-(A::AbstractTensor{rank}, B::AbstractTensor{rank}) where rank

Binary minus. Creates an AddOperation of A and -1*B.

Base.:/ — Method

Base.:/(A::AbstractTensor, y::AbstractTensor{0})

Creates a DivisionOperation where A is divived by y. y must be scalar-shaped.

Base.:/ — Method

Base.:/(A::AbstractTensor, y::Number)

Creates a DivisionOperation where A is divived by y.

Base.:/ — Method

Base.:/(y::Number, A::AbstractTensor{0})

Creates a DivisionOperation where y is divided by A, where A must be a scalar.

Base.:^ — Method

Base.:^(x::AbstractTensor{0}, y::AbstractTensor{0})

Raises the scalar-shaped tensor x to the scalar-shaped power y. Returns a PowerOperation.

Base.:^ — Method

Base.:^(x::AbstractTensor{0}, y::Number)

Raises the scalar-shaped tensor x to the power y. Returns a PowerOperation.

Base.getindex — Method

Base.getindex(A::AbstractTensor, ys::Vararg{Index})

Creates an IndexingOperation of A indexed by ys. If every dimension of A is not indexed, creates a ComponentTensorOperation over the unindexed dimensions.

Base.getindex — Method

Base.getindex(A::AbstractTensor, ys::Vararg{Union{String, Int, Index}})

A convenience function that allows indexing a tensor by an integer. The function creates a corresponding FixedIndex object in the appropiate vector space.

Technically the function can also create a FreeIndex from a string, but this is not recommended, because the syntax makes no distinction between upper and lower indices.

Base.sqrt — Method

Base.sqrt(x::AbstractTensor{0})

Returns the square root of x as a PowerOperation.

Base.transpose — Method

Base.transpose(A::AbstractTensor)

Reverses the shape of A. Does this by indexing A and then creating a ComponentTensorOperation with the indices reversed.

TensorDSL.:⊗ — Method

⊗(A::AbstractTensor, B::AbstractTensor)

Returns an OuterProductOperation with A and B as its children.

TensorDSL.componenttensor — Method

componenttensor(A::AbstractTensor, indices::Vararg{Index})

Creates a ComponentTensorOperation of A over indices. Indices must be a subset of the free indices of A.

TensorDSL.contractioncheck — Method

contractioncheck(A::AbstractTensor)

Checks if we have have an upper and lower index. If there is, we create an IndexSumOperation node.

TensorDSL.trace — Method

trace(A::AbstractTensor{2})

Returns the trace of a matrix of the shape (V, V').

Trigonometry

Base.acos — Method

Base.acos(s::AbstractTensor{0})

Returns a ArccosineOperation represeting arccosine of s.

Base.asin — Method

Base.asin(s::AbstractTensor{0})

Returns a ArcsineOperation represeting arcsine of s.

Base.atan — Method

Base.atan(s::AbstractTensor{0})

Returns a ArctangentOperation represeting arctangent of s.

Base.cos — Method

Base.cos(s::AbstractTensor{0})

Returns a CosineOperation represeting cosine of s in radians.

Base.sin — Method

Base.sin(s::AbstractTensor{0})

Returns a SineOperation represeting sine of s in radians.

Base.tan — Method

Base.tan(s::AbstractTensor{0})

Returns a TangentOperation represeting tangent of s in radians.

TensorDSL.ArccosineOperation — Type

CosineOperation(A::AbstractTensor{rank}) where rank

Represents computing cos of A.

TensorDSL.ArcsineOperation — Type

ArcsineOperation(A::AbstractTensor{rank}) where rank

Represents computing arcsin of A.

TensorDSL.ArctangentOperation — Type

ArctangentOperation(A::AbstractTensor{rank}) where rank

Represents computing arctan of A.

TensorDSL.CosineOperation — Type

CosineOperation(A::AbstractTensor{rank}) where rank

Represents computing cos of A.

TensorDSL.SineOperation — Type

SineOperation(A::AbstractTensor{rank}) where rank

Represents computing sin of A.

TensorDSL.TangentOperation — Type

TangentOperation(A::AbstractTensor{rank}) where rank

Represents computing tan of A.

Differentation

TensorDSL.differentiate — Method

differentiate(A::AbstractTensor{0}, x::VariableTensor{0})

Differentiates A w.r.t x.

TensorDSL.divergence — Method

divergence(A::AbstractTensor{1}, vars::Vararg{VariableTensor{0}})

Computes the divergence of a vector-shaped tensor A w.r.t vars.

Operation nodes

TensorDSL.AddOperation — Type

AddOperation(children::Tuple{Vararg{AbstractTensor{rank}}},
    freeindices::Tuple{Vararg{FreeIndex}}) where rank

A node symbolising addition of tensors.

TensorDSL.ComponentTensorOperation — Type

ComponentTensorOperation(shape::Tuple{Vararg{AbstractVectorSpace}},
    children::Tuple{AbstractTensor, Indices},
    freeindices::Tuple{Vararg{FreeIndex}})

A node symbolising component tensor.

TensorDSL.DivisionOperation — Type

DivisionOperation(A::AbstractTensor, y::AbstractTensor{0})

Symbolises dividing tensor A by a scalar-shaped tensor y.

TensorDSL.IndexSumOperation — Type

IndexSumOperation(A::AbstractTensor, indices::Indices, freeindices::Vararg{FreeIndex})

Symbolises summing over given free index of a tensor.

TensorDSL.IndexingOperation — Type

IndexingOperation(x::AbstractTensor, indices::Indices)

A node symbolising indexing a tensor by its every dimension.

TensorDSL.Operation — Type

Operation{rank} <: AbstractTensor{rank}

A supertype for all tensor operations.

TensorDSL.OuterProductOperation — Type

OuterProductOperation(shape::Tuple{Vararg{AbstractVectorSpace}},
    children::Tuple{AbstractTensor, AbstractTensor},

A node symbolising tensor outer product.

TensorDSL.PowerOperation — Type

PowerOperation(x::AbstractTensor{0}, y::AbstractTensor{0})

Symbolises raising a scalar-shaped tensor x to a scalar-shaped power y.