TensorDSL.AbstractTensor — Type

A type for all the tensors and tensor like objects.

The rank means the number of indices. For example AbstractTensor{0} is a scalar, and AbstractTensor{2} a matrix.

TensorDSL.TerminalTensor — Type

A type for the tensors which are terminal nodes, i.e not created by operations.

TensorDSL.Scalar — Constant

A union type for everything that can be treated as a scalar.

TensorDSL.VariableTensor — Type

Tensors of which we only know in which vector spaces their indices are.

TensorDSL.Tensor — Type

Tensor created from a multidimensional array.

TensorDSL.ZeroTensor — Type

Symbolic zero tensor.

TensorDSL.ConstantTensor — Type

A tensor where every entry is of the same value.